able to build phase 3

+++ src/user.py
@@ -24,39 +24,36 @@ def createLfs():
     groupadd lfs 
     useradd -s /bin/bash -g lfs -m -k /dev/null lfs
     """
-    try:
-        log_file = Path("logs/001_creating_user.log")
-        log_file.parent.mkdir(parents=True, exist_ok=True)
-        
-        result = subprocess.run(
-            ["bash"],
-            input=input_command,
-            text=True,
-            check=False,
-            capture_output=True,
-        )
-        
-        with log_file.open("w") as f:
-            f.write("=== STDOUT ===\n")
-            f.write(result.stdout)
-            f.write("== STDERR ===\n")
-            f.write(result.stderr)
-            f.close()
-        
-        if result.returncode == 0:
-            ConsoleMSG.passed("Created user and group lfs")
-            return True
-        else:
-            ConsoleMSG.failed("Could not create user and group lfs")
-            return False
-    except KeyError:
-        ConsoleMSG.failed(f"Could not create lfs user and group {KeyError}")
-        return False        
+    if (doesLfsExist() == False):
+        try:
+            log_file = Path("logs/001_creating_user.log")
+            log_file.parent.mkdir(parents=True, exist_ok=True)
             
+            result = subprocess.run(
+                ["bash"],
+                input=input_command,
+                text=True,
+                check=False,
+                capture_output=True,
+            )
             
-
-#    groupadd lfs
-#useradd -s /bin/bash -g lfs -m -k /dev/null lfs
+            with log_file.open("w") as f:
+                f.write("=== STDOUT ===\n")
+                f.write(result.stdout)
+                f.write("== STDERR ===\n")
+                f.write(result.stderr)
+                f.close()
+            
+            if result.returncode == 0:
+                ConsoleMSG.passed("Created user and group lfs")
+                return True
+            else:
+                ConsoleMSG.failed("Could not create user and group lfs")
+                return False
+        except KeyError:
+            ConsoleMSG.failed(f"Could not create lfs user and group {KeyError}")
+            return False        
+            
 
 # delete lfs account
 def deleteLfs():
@@ -64,35 +61,35 @@ def deleteLfs():
     groupdel lfs 
     userdel -r lfs
     """
-    
-    try:
-        log_file = Path("logs/001_creating_user.log")
-        log_file.parent.mkdir(parents=True, exist_ok=True)
-        
-        result = subprocess.run(
-            ["bash"],
-            input=input_command,
-            text=True,
-            check=False,
-            capture_output=True,
-        )
-        
-        with log_file.open("w") as f:
-            f.write("=== STDOUT ===\n")
-            f.write(result.stdout)
-            f.write("== STDERR ===\n")
-            f.write(result.stderr)
-            f.close()
-        
-        if result.returncode == 0:
-            ConsoleMSG.passed("Created lfs user and group")
-            return True
-        else:
-            ConsoleMSG.failed("Could not create lfs user and group")
-            return False
-    except KeyError:
-        ConsoleMSG.failed(f"Could not create lfs user and group {KeyError}")
-        return False  
+    if (doesLfsExist() == True ):
+        try:
+            log_file = Path("logs/001_creating_user.log")
+            log_file.parent.mkdir(parents=True, exist_ok=True)
+            
+            result = subprocess.run(
+                ["bash"],
+                input=input_command,
+                text=True,
+                check=False,
+                capture_output=True,
+            )
+            
+            with log_file.open("w") as f:
+                f.write("=== STDOUT ===\n")
+                f.write(result.stdout)
+                f.write("== STDERR ===\n")
+                f.write(result.stderr)
+                f.close()
+            
+            if result.returncode == 0:
+                ConsoleMSG.passed("Removed lfs user and group")
+                return True
+            else:
+                ConsoleMSG.failed("Could not remove lfs user and group")
+                return False
+        except KeyError:
+            ConsoleMSG.failed(f"Could not remove lfs user and group {KeyError}")
+            return False  
     
 def chownBuildDir(config: GlobalConfig):
     input_command = "chown -Rv lfs:lfs " + str(config.build_path) + "\n chmod -R u+rwX " + str(config.build_path)
